# docs/swagger/generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Convo API",
    "version": "1.0.0",
    "description": "\n      **Convo API Documentation**\n      \n      A comprehensive API for managing conversational forms, workspaces, and user interactions.\n      \n      ## Authentication\n      This API uses Clerk for authentication. Include your session token in the Authorization header:\n      `Authorization: Bearer <your-token>`\n      \n      ## Rate Limiting\n      API requests are rate limited per user:\n      - **Free Plan**: 100 requests/hour\n      - **Pro Plan**: 1000 requests/hour\n      - **Enterprise**: Custom limits\n      \n      Rate limit headers are included in all responses:\n      - `X-RateLimit-Limit`: Maximum requests allowed\n      - `X-RateLimit-Remaining`: Requests remaining in current window\n      - `X-RateLimit-Reset`: Time when the limit resets (Unix timestamp)\n    ",
    "contact": {
      "name": "API Support",
      "email": "support@convo.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-generated-at": "2025-07-01T10:52:11.739Z",
    "x-generator": "convo-forms-swagger-generator"
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    },
    {
      "url": "https://convo.ai/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ClerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token for user authentication"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for programmatic access"
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "message": {
                  "type": "string",
                  "example": "Authentication required to access this resource"
                },
                "code": {
                  "type": "string",
                  "example": "AUTH_REQUIRED"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Access forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                },
                "message": {
                  "type": "string",
                  "example": "Insufficient permissions to access this resource"
                },
                "code": {
                  "type": "string",
                  "example": "INSUFFICIENT_PERMISSIONS"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Not Found"
                },
                "message": {
                  "type": "string",
                  "example": "The requested resource was not found"
                },
                "code": {
                  "type": "string",
                  "example": "NOT_FOUND"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Validation Error"
                },
                "message": {
                  "type": "string",
                  "example": "Request validation failed"
                },
                "code": {
                  "type": "string",
                  "example": "VALIDATION_ERROR"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string",
                        "example": "email"
                      },
                      "message": {
                        "type": "string",
                        "example": "Invalid email format"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Internal Server Error"
                },
                "message": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "Form": {
        "type": "object",
        "description": "Form definition with fields and settings",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique form identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid",
            "description": "Parent workspace ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "createdBy": {
            "type": "string",
            "description": "Creator's user ID",
            "example": "user_2NiWrEwuDBCQ7XiZpFxGD7CpYKx"
          },
          "title": {
            "type": "string",
            "description": "Form title",
            "example": "Customer Feedback Survey",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Form description",
            "example": "Help us improve our service by sharing your feedback",
            "maxLength": 1000
          },
          "prompt": {
            "type": "string",
            "description": "AI generation prompt used to create this form",
            "example": "Create a customer feedback form with rating and comment fields"
          },
          "isConversational": {
            "type": "boolean",
            "description": "Whether form uses conversational mode",
            "example": false,
            "default": false
          },
          "isPublished": {
            "type": "boolean",
            "description": "Whether form is published and accepting responses",
            "example": true,
            "default": false
          },
          "config": {
            "$ref": "#/components/schemas/FormConfig"
          },
          "settings": {
            "$ref": "#/components/schemas/FormSettings"
          },
          "version": {
            "type": "integer",
            "description": "Form version number",
            "example": 1,
            "default": 1
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When form was published",
            "example": "2024-01-15T12:00:00Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Form creation timestamp",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last form update timestamp",
            "example": "2024-01-20T14:45:00Z"
          },
          "responseCount": {
            "type": "integer",
            "description": "Total number of responses (computed field)",
            "example": 142,
            "readOnly": true
          }
        },
        "required": ["id", "workspaceId", "createdBy", "title", "config", "createdAt", "updatedAt"]
      },
      "FormConfig": {
        "type": "object",
        "description": "Form structure and field definitions",
        "properties": {
          "fields": {
            "type": "array",
            "description": "Array of form fields",
            "items": {
              "$ref": "#/components/schemas/FormField"
            }
          },
          "settings": {
            "$ref": "#/components/schemas/FormDisplaySettings"
          },
          "styling": {
            "$ref": "#/components/schemas/FormStyling"
          }
        },
        "required": ["fields"]
      },
      "FormField": {
        "type": "object",
        "description": "Individual form field definition",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique field identifier",
            "example": "field_1"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "email",
              "phone",
              "select",
              "multiselect",
              "textarea",
              "number",
              "date",
              "file",
              "signature",
              "rating"
            ],
            "description": "Field input type",
            "example": "text"
          },
          "label": {
            "type": "string",
            "description": "Field label displayed to users",
            "example": "Full Name",
            "maxLength": 200
          },
          "placeholder": {
            "type": "string",
            "description": "Placeholder text",
            "example": "Enter your full name",
            "maxLength": 200
          },
          "required": {
            "type": "boolean",
            "description": "Whether field is required",
            "example": true,
            "default": false
          },
          "order": {
            "type": "integer",
            "description": "Field display order",
            "example": 1,
            "minimum": 0
          },
          "options": {
            "type": "array",
            "description": "Options for select/multiselect fields",
            "items": {
              "type": "string"
            },
            "example": ["Option 1", "Option 2", "Option 3"]
          },
          "validation": {
            "$ref": "#/components/schemas/FieldValidation"
          }
        },
        "required": ["id", "type", "label", "required", "order"]
      },
      "FieldValidation": {
        "type": "object",
        "description": "Field validation rules",
        "properties": {
          "minLength": {
            "type": "integer",
            "description": "Minimum character length",
            "example": 2,
            "minimum": 0
          },
          "maxLength": {
            "type": "integer",
            "description": "Maximum character length",
            "example": 100,
            "minimum": 1
          },
          "pattern": {
            "type": "string",
            "description": "RegEx validation pattern",
            "example": "^[A-Za-z\\s]+$"
          },
          "min": {
            "type": "number",
            "description": "Minimum numeric value",
            "example": 0
          },
          "max": {
            "type": "number",
            "description": "Maximum numeric value",
            "example": 100
          }
        }
      },
      "FormDisplaySettings": {
        "type": "object",
        "description": "Form display and behavior settings",
        "properties": {
          "submitText": {
            "type": "string",
            "description": "Custom submit button text",
            "example": "Send Feedback",
            "default": "Submit"
          },
          "redirectUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to redirect after submission",
            "example": "https://example.com/thank-you"
          },
          "emailNotifications": {
            "type": "boolean",
            "description": "Send email notifications for responses",
            "example": true,
            "default": false
          },
          "collectEmail": {
            "type": "boolean",
            "description": "Automatically collect submitter email",
            "example": true,
            "default": false
          },
          "requireAuth": {
            "type": "boolean",
            "description": "Require user authentication to submit",
            "example": false,
            "default": false
          }
        }
      },
      "FormStyling": {
        "type": "object",
        "description": "Form visual customization",
        "properties": {
          "backgroundColor": {
            "type": "string",
            "description": "Background color (hex)",
            "example": "#ffffff",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "textColor": {
            "type": "string",
            "description": "Text color (hex)",
            "example": "#333333",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "buttonColor": {
            "type": "string",
            "description": "Button color (hex)",
            "example": "#007bff",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "fontFamily": {
            "type": "string",
            "description": "Font family",
            "example": "Inter, sans-serif"
          }
        }
      },
      "FormSettings": {
        "type": "object",
        "description": "Form behavior and integration settings",
        "properties": {
          "allowMultipleSubmissions": {
            "type": "boolean",
            "description": "Allow multiple submissions from same user",
            "example": false,
            "default": true
          },
          "collectAnalytics": {
            "type": "boolean",
            "description": "Collect analytics data",
            "example": true,
            "default": true
          },
          "enableCaptcha": {
            "type": "boolean",
            "description": "Enable CAPTCHA protection",
            "example": false,
            "default": false
          },
          "webhookUrl": {
            "type": "string",
            "format": "uri",
            "description": "Webhook URL for response notifications",
            "example": "https://api.example.com/webhook"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/templates/{id}/create-form": {
      "post": {
        "summary": "Create a form from template",
        "description": "Creates a new form using the specified template as a starting point.\nCopies the template's form schema and increments the template's usage count.\nRequires create_form permission in the target workspace.\n",
        "tags": ["Templates"],
        "security": [
          {
            "ClerkAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Template ID to create form from"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["workspaceId", "name"],
                "properties": {
                  "workspaceId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Target workspace ID for the new form"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Name for the new form"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description for the new form (optional)"
                  },
                  "isConversational": {
                    "type": "boolean",
                    "default": false,
                    "description": "Whether to enable conversational mode"
                  },
                  "isPublished": {
                    "type": "boolean",
                    "default": false,
                    "description": "Whether to publish the form immediately"
                  }
                }
              },
              "examples": {
                "createFormFromTemplate": {
                  "summary": "Create form from job application template",
                  "value": {
                    "workspaceId": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Software Engineer Application",
                    "description": "Application form for software engineering positions",
                    "isConversational": true,
                    "isPublished": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Form created successfully from template",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "form": {
                      "$ref": "#/components/schemas/Form"
                    },
                    "template": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Form created successfully from template"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}